fix: change sort_order from INTEGER to BIGINT to support Date.now() values

+++ supabase-schema.sql
@@ -3,7 +3,7 @@
 -- Run this in your Supabase SQL Editor
 -- ============================================
 
--- Enable required extensions
+-- Enable required extenshions
 CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
 
 -- ============================================
@@ -51,7 +51,7 @@ CREATE TABLE items (
   category TEXT DEFAULT 'general',
   barcode TEXT,
   created_at TIMESTAMPTZ DEFAULT NOW(),
-  sort_order INTEGER NOT NULL DEFAULT 0,
+  sort_order BIGINT NOT NULL DEFAULT 0,
   updated_at TIMESTAMPTZ DEFAULT NOW()
 );
 
@@ -62,7 +62,7 @@ CREATE INDEX idx_items_store ON items(store);
 CREATE INDEX idx_items_sort ON items(household_id, week_id, store, sort_order);
 
 -- Idempotent migration for existing deployments
-ALTER TABLE items ADD COLUMN IF NOT EXISTS sort_order INTEGER NOT NULL DEFAULT 0;
+ALTER TABLE items ADD COLUMN IF NOT EXISTS sort_order BIGINT NOT NULL DEFAULT 0;
 
 -- ============================================
 -- 4. PRICE MEMORY
